add new TransferWaitingPage
