+++ app/components/transfer/TransferWaitingPage.scss
@@ -1,0 +1,3 @@
+
+.component {
+}
